+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `FetchInstruction.placeholder_attributes` property + `AbstractFetcher` handling.
 - Add SQLAlchemy ORM fetcher example.
 - Add the `add_skip_file_prefix()` function to control warning stack levels.
+- Add `translator_typing.GoOfflineParams` typed dict.
+- Add `Translator.go_offline(raise_if_offline=False)`
 
 ### Changed
 - Links in warnings and exceptions now to the current version in the docs (instead of the _'stable'_ version).